Recap of week one activities
Week one recap @ TRYBE- X
#trybe-X #trybe-xBootcamp 2024 @Trube-X Ibrahim abdulquddus Fattahandi @musamuhd25188

what I learned during trybe-x Bootcamp are as follows
Day 1) introduction to Internet

Day 2) Introduction to web technology

Day 3) setting up development Environment

Day 4) Basic Command line

Day 5) introduction to web Development work flow

I will discuss on few things I learned from day one and two.
Day 1 : introduction to Internet.

Internet is a place of information, entertainment, and transfer of information.it is the largest computer network connecting billions of user from different countries. I used to think Internet is just to buy data in phone and check my phone network whether they is network or not but computer is more than that,it is a network of network.
As we all know, Internet connect computer all together and this can be achieved by the used of a router(it works like a railway station).
Device that forward data among each other is called packet switches.
Internet are governing by protocol,such as the TCP(Transmission control protocol)
Http(Hyper text transfer protocol
DNS (Domain name system) etc.
There are two types of network which are:the intranet and the extranet
Intranet are private networks that accessed by a particular organization or are restricted to members of a particular organization
Extranet are open to all or part of a private networks to allow sharing and collaboration with company and organization
I also learned what URL means and the used
URL means Uniform resource locator
URL is an address of a unique resource on the Internet or it is a specific address of resource on the internet.it is used by the browser to retrieve published resources such as Css,pages and documents etc
There are two types of URL
Which is absolute URL and Relative URL
Absolute URL contains all the necessary parts needed to locate the resources.
Relative URL contains some part but some part are missing like Scheme or Domain or sub-resource etc.
Day2 was introduction to web technologies

Web literally mean to connect complex things together
Computer connected to the Internet are called client and server.
Client is a typical connected device
While
Server are computer that store webpages,site or apps etc.
I also learned what it means my HTML basic
From the beginning you will be seeing acronym, yeah that is how it is in computer to make writing to be short and elaborate.
HTML means hyper text mark up language. It is a code that is used to structure a web page and it content.for example, content could be structured within a set of paragraphs a list of points,or using images and data tables
There are different steps to consider before coding with HTML
The steps are opening tag, closing tag and content
Opening tag is like this<p>
Closing tag is like this </p>
Content: is the text.
Element: opening tag, closing tag, and makeup element
<h1></h1> it means Head one to make something to be bold.
There are many things I learned for the first weeks which can make someone understand the basics of web development,and I urge my readers to please follow #trybe_X if want to learn web development and if you have knowledge of web development you can still follow their daily lesson to gain more knowledge.
I can’t wait to get more topics to learn from in their next class,.
Thank you very much Abodulqudu for this great privilege.
#trybe_X Bootcap 2024